feat: добавить авторизацию и мидлверы

+++ controllers/cards.js
@@ -1,58 +1,65 @@
 const Card = require('../models/card');
+const BadRequestError = require('../errors/bad-request-error');
+const NotFoundError = require('../errors/not-found-error');
+const ServerError = require('../errors/server-error');
+const ForbiddenError = require('../errors/forbidden-error');
 const {
   statusCodeOK,
   statusCodeCreate,
-  statusCodeBadRequest,
-  statusCodeNotFound,
-  statusCodeServerError,
 } = require('../constants/statusCodeConstatns');
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.status(statusCodeOK).send(cards))
-    .catch(() => res.status(statusCodeServerError).send({ message: 'Server error' }));
+    .catch(() => {
+      const err = new ServerError('Server Error');
+      next(err);
+    });
 };
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   req.body.owner = req.user._id;
   const newCardData = req.body;
   Card.create(newCardData)
     .then((newCard) => res.status(statusCodeCreate).send(newCard))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        return res
-          .status(statusCodeBadRequest)
-          .send({ message: 'Incorrect data was transmitted' });
+    .catch((e) => {
+      if (e.name === 'ValidationError') {
+        const err = new BadRequestError('Incorrect data was transmitted');
+        next(err);
       }
-      return res
-        .status(statusCodeServerError)
-        .send({ message: 'Server error' });
+      const err = new ServerError('Server Error');
+      next(err);
     });
 };
 
-const removeCard = (req, res) => {
+const removeCard = (req, res, next) => {
   const { cardId } = req.params;
-  Card.findByIdAndDelete(cardId)
+  Card.findById(cardId)
     .then((card) => {
       if (!card) {
-        return res
-          .status(statusCodeNotFound)
-          .send({ message: 'The requested information was not found' });
-      } return res.status(statusCodeOK).send({ message: 'done' });
-    })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res
-          .status(statusCodeBadRequest)
-          .send({ message: 'Incorrect data was transmitted' });
+        const err = new NotFoundError('The requested information was not found');
+        next(err);
+        return;
       }
-      return res
-        .status(statusCodeServerError)
-        .send({ message: 'Server error' });
+      if (card.owner.toString() !== req.user._id) {
+        const err = new ForbiddenError('Access is restricted');
+        next(err);
+        return;
+      }
+      Card.findByIdAndDelete(cardId)
+        .then(() => res.status(statusCodeOK).send({ message: 'done' }))
+        .catch((e) => {
+          if (e.name === 'CastError') {
+            const err = new BadRequestError('Incorrect data was transmitted');
+            next(err);
+          }
+          const err = new ServerError('Server Error');
+          next(err);
+        });
     });
 };
 
-const likeCard = (req, res) => {
+const likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -60,24 +67,22 @@ const likeCard = (req, res) => {
   )
     .then((updateCard) => {
       if (!updateCard) {
-        return res
-          .status(statusCodeNotFound)
-          .send({ message: 'The requested information was not found' });
-      } return res.status(statusCodeOK).send(updateCard);
+        const err = new NotFoundError('The requested information was not found');
+        next(err);
+        return;
+      } res.status(statusCodeOK).send(updateCard);
     })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res
-          .status(statusCodeBadRequest)
-          .send({ message: 'Incorrect data was transmitted' });
+    .catch((e) => {
+      if (e.name === 'CastError') {
+        const err = new BadRequestError('Incorrect data was transmitted');
+        next(err);
       }
-      return res
-        .status(statusCodeServerError)
-        .send({ message: 'Server error' });
+      const err = new ServerError('Server Error');
+      next(err);
     });
 };
 
-const unlikeCard = (req, res) => {
+const unlikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -85,20 +90,18 @@ const unlikeCard = (req, res) => {
   )
     .then((updateCard) => {
       if (!updateCard) {
-        return res
-          .status(statusCodeNotFound)
-          .send({ message: 'The requested information was not found' });
-      } return res.status(statusCodeOK).send(updateCard);
+        const err = new NotFoundError('The requested information was not found');
+        next(err);
+        return;
+      } res.status(statusCodeOK).send(updateCard);
     })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res
-          .status(statusCodeBadRequest)
-          .send({ message: 'Incorrect data was transmitted' });
+    .catch((e) => {
+      if (e.name === 'CastError') {
+        const err = new BadRequestError('Incorrect data was transmitted');
+        next(err);
       }
-      return res
-        .status(statusCodeServerError)
-        .send({ message: 'Server error' });
+      const err = new ServerError('Server Error');
+      next(err);
     });
 };
